+++ .vscode/settings.json
@@ -1,8 +1,6 @@
 {
   "python.testing.pytestEnabled": true,
-  "python.testing.pytestArgs": [
-    "tests"
-  ],
+  "python.testing.pytestArgs": ["tests"],
   "ruff.importStrategy": "fromEnvironment",
   "mypy-type-checker.importStrategy": "fromEnvironment",
   "[python]": {
@@ -13,5 +11,9 @@
       "source.organizeImports.ruff": "explicit"
     }
   },
-  "workbench.colorCustomizations": { /* do not change please... */}
+  "workbench.colorCustomizations": {
+    /* do not change please... */
+  },
+  "python-envs.defaultEnvManager": "ms-python.python:venv",
+  "python-envs.defaultPackageManager": "ms-python.python:pip"
 }
